B-231: M-X64-S1 — syslinux.c compiles clean; make bootsbl succeeds
Fix x64 port: mword/muword, MINIMAL_* macros, missing symbols, typet label.

osint/port.h:
- typedef mword=long, muword=ulong (x64: word=long=8 bytes)
- #include <stdint.h> for uintptr_t
- MINFRAME=0, ARGPUSHSIZE=0, SA(n) for System V AMD64 ABI

osint/extern32.h:
- mword int->long, muword uint->ulong (32-bit leftovers)

osint/osint.h:
- MINIMAL_ALLOC/ALOCS/ALOST/BLKLN uppercase aliases for enum
- TYPET alias for typet data symbol
- B_EFC alias for b_efc code symbol

osint/sproto.h:
- callef/loadef: word->mword (matching syslinux.c definitions)

osint/syslinux.c:
- static initsels: implicit int -> explicit int
- trig externs: double(double) -> void(void) matching math.c

osint/sysex.c:
- MP_OFF: fix 1-arg call to 2-arg macro

osint/syslinux_float.c (new):
- f_2_i, i_2_f, f_add, f_sub, f_mul, f_div, f_neg: C-ABI bridge
  for APDF table (missing from math.c which uses reg_ra convention)

int.asm:
- pushregs/popregs: aliases -> save_regs/restore_regs
- callextfun: x64 trampoline for LOAD'd external functions

bootstrap/sbl.asm + sbl.asm:
- typet: label added before type table data (global declared but never labelled)

+++ osint/extern32.h
@@ -25,8 +25,8 @@ Copyright 2012-2017 David Shields
  */
 
 # include "system.h"
-typedef int mword;           /* MINIMAL word */
-typedef unsigned int muword; /* MINIMAL unsigned word */
+typedef long mword;           /* MINIMAL word — x64: long = 8 bytes */
+typedef unsigned long muword; /* MINIMAL unsigned word — x64: unsigned long = 8 bytes */
 # ifndef far
 #  define far
 # endif
